+++ src/imio/smartweb/core/viewlets/banner.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 
 from imio.smartweb.core.browser.banner.settings import ILocallyHiddenBanner
+from imio.smartweb.core.utils import hash_md5
 from plone import api
 from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.layout.viewlets import common
 from plone.namedfile.file import NamedBlobImage
 from Products.CMFPlone.utils import base_hasattr
-from zope.component import queryMultiAdapter
 
 
 class BannerViewlet(common.ViewletBase):
@@ -54,8 +54,9 @@ class BannerViewlet(common.ViewletBase):
     def background_style(self):
         if self.is_banner_hidden:
             return ""
-        images_view = queryMultiAdapter((self.banner_item, self.request), name="images")
-        scale = images_view.scale("banner", "banner")
-        css_bg_image = f"background-image:url({scale.url});"
+        modification_date = self.banner_item.ModificationDate()
+        modified_hash = hash_md5(modification_date)
+        image_url = f"{self.banner_item.absolute_url()}/@@images/banner?cache_key={modified_hash}"
+        css_bg_image = f"background-image:url({image_url});"
         css_bg_size = "background-size:cover;"
         return " ".join([css_bg_image, css_bg_size])